# WIP: avoid allocations in inplace methods (proof of concept)