--- a/src/TaylorSeries.jl
+++ b/src/TaylorSeries.jl
@@ -16,7 +16,7 @@ see also [`HomogeneousPolynomial`](@ref).
 """
 module TaylorSeries
 
-
+using InteractiveUtils
 using SparseArrays: SparseMatrixCSC
 using Markdown