--- a/src/arithmetic.jl
+++ b/src/arithmetic.jl
@@ -203,6 +203,30 @@ end
 ones(::Type{HomogeneousPolynomial{T}}, order::Int) where {T<:Number} =
     ones( HomogeneousPolynomial([one(T)], 0), order)
 
+function zero!(a::Taylor1{T}) where {T<:NumberNotSeries}
+    a.coeffs .= zero.(a.coeffs)
+    return nothing
+end
+
+function zero!(a::HomogeneousPolynomial{T}) where {T<:NumberNotSeries}
+    a.coeffs .= zero.(a.coeffs)
+    return nothing
+end
+
+function zero!(a::TaylorN{T}) where {T<:NumberNotSeries}
+    for i in 0:a.order
+        zero!(a[i])
+    end
+    return nothing
+end
+
+function zero!(a::Taylor1{TaylorN{T}}) where {T<:NumberNotSeries}
+    for i in 0:a.order
+        zero!(a[i])
+    end
+    return nothing
+end
+
 
 
 ## Addition and subtraction ##
@@ -479,10 +503,12 @@ for T in (:Taylor1, :TaylorN)
     end
 end
 
+########
+
 function mul!(res::Taylor1{TaylorN{T}}, a::Taylor1{TaylorN{T}}, b::Taylor1{TaylorN{T}},
         ordT::Int) where {T<:NumberNotSeries}
     # Sanity
-    zero!(res, a, ordT)
+    zero!(res[ordT])
     for k in 0:ordT
         @inbounds for ordQ in eachindex(a[ordT])
             mul!(res[ordT], a[k], b[ordT-k], ordQ)