--- a/src/arithmetic.jl
+++ b/src/arithmetic.jl
@@ -746,6 +746,7 @@ end
 
 @inline function div!(c::Taylor1, a::NumberNotSeries,
         b::Taylor1, k::Int)
+    ### TODO: use non-allocating div!(::TaylorN,::TaylorN,::TaylorN)
     iszero(a) && !iszero(b) && zero!(c, k)
     # order and coefficient of first factorized term
     # In this case, since a[k]=0 for k>0, we can simplify to:
@@ -835,26 +836,6 @@ mul!(res::Taylor1{TaylorN{T}}, a::Taylor1{TaylorN{T}},
     end
     return nothing
 end
-
-# ### TODO: use non-allocating div!(::TaylorN,::TaylorN,::TaylorN)
-# @inline function div!(c::Taylor1{TaylorN{T}}, a::NumberNotSeries,
-#         b::Taylor1{TaylorN{T}}, k::Int) where {T<:NumberNotSeries}
-#     # order and coefficient of first factorized term
-#     # In this case, since a[k]=0 for k>0, we can simplify to:
-#     # ordfact, cdivfact = 0, a/b[0]
-#     if k == 0
-#         @inbounds c[0] = a/b[0]
-#         return nothing
-#     end
-
-#     imin = max(0, k-b.order)
-#     @inbounds c[k] = c[imin] * b[k-imin]
-#     @inbounds for i = imin+1:k-1
-#         c[k] += c[i] * b[k-i]
-#     end
-#     @inbounds c[k] = -c[k]/b[0]
-#     return nothing
-# end