--- a/src/power.jl
+++ b/src/power.jl
@@ -276,7 +276,7 @@ end
     end
 
     # Sanity
-    zero!(res[ordT])
+    zero!(res, ordT)
 
     # First non-zero coefficient
     l0 = findfirst(a)
@@ -311,6 +311,7 @@ end
     end
     @inbounds for ordQ in eachindex(tmp)
         tmp1[ordQ] = tmp[ordQ]/kprime
+        @show @which div!(res[ordT], tmp1, a[l0], ordQ)
         div!(res[ordT], tmp1, a[l0], ordQ)
     end
 
@@ -411,7 +412,7 @@ end
 @inline function sqr!(res::Taylor1{TaylorN{T}}, a::Taylor1{TaylorN{T}},
         ordT::Int) where {T<:NumberNotSeries}
     # Sanity
-    zero!(res[ordT])
+    zero!(res, ordT)
     if ordT == 0
         @inbounds for ordQ in eachindex(a[0])
             @inbounds sqr!(res[0], a[0], ordQ)
@@ -628,7 +629,7 @@ end
 @inline function sqrt!(res::Taylor1{TaylorN{T}}, a::Taylor1{TaylorN{T}}, ordT::Int,
         ordT0::Int=0) where {T<:NumberNotSeries}
     # Sanity
-    zero!(res[ordT])
+    zero!(res, ordT)
     ordT < ordT0 && return nothing
 
     if ordT == ordT0
@@ -656,6 +657,7 @@ end
         end
     end
 
+    @show res
     @inbounds for ordQ in eachindex(a[0])
         tmp[ordQ] = -2 * tmp[ordQ]
         if ordT+ordT0 ≤ a.order
@@ -668,6 +670,7 @@ end
         tmp1[ordQ] = 2*res[ordT0][ordQ]
         div!(res[ordT], tmp, tmp1, ordQ)
     end
+    @show res
 
     return nothing
 end