In current `master`, inplace methods allocate, especially for `TaylorN` and mixtures:

```julia
julia> using TaylorSeries, BenchmarkTools

julia> dq = set_variables("dq", order=2, numvars=2)
2-element Vector{TaylorN{Float64}}:
  1.0 dq₁ + 𝒪(‖x‖³)
  1.0 dq₂ + 𝒪(‖x‖³)

julia> x1N = Taylor1([dq[1],dq[2],dq[1]*dq[2]])
  1.0 dq₁ + 𝒪(‖x‖³) + ( 1.0 dq₂ + 𝒪(‖x‖³)) t + ( 1.0 dq₁ dq₂ + 𝒪(‖x‖³)) t² + 𝒪(t³)

julia> y1N = atan(x1N)
  1.0 dq₁ + 𝒪(‖x‖³) + ( 1.0 dq₂ + 𝒪(‖x‖³)) t + ( 1.0 dq₁ dq₂ + 𝒪(‖x‖³)) t² + 𝒪(t³)

julia> z1N = zero(x1N)
  0.0 + 𝒪(‖x‖³) + 𝒪(t³)

julia> @btime TaylorSeries.mul!($z1N,$x1N,$y1N,0)
  590.044 ns (14 allocations: 1.06 KiB)
```

This PR is a proof of concept which avoids allocations in key places, such as `zero` and `mul!` (similar things can be tried in `div!`, etc.), just by defining suitable inplace `zero!` methods:

```julia
julia> using TaylorSeries, BenchmarkTools

julia> dq = set_variables("dq", order=2, numvars=2)
2-element Vector{TaylorN{Float64}}:
  1.0 dq₁ + 𝒪(‖x‖³)
  1.0 dq₂ + 𝒪(‖x‖³)

julia> x1N = Taylor1([dq[1],dq[2],dq[1]*dq[2]])
  1.0 dq₁ + 𝒪(‖x‖³) + ( 1.0 dq₂ + 𝒪(‖x‖³)) t + ( 1.0 dq₁ dq₂ + 𝒪(‖x‖³)) t² + 𝒪(t³)

julia> y1N = atan(x1N)
  1.0 dq₁ + 𝒪(‖x‖³) + ( 1.0 dq₂ + 𝒪(‖x‖³)) t + ( 1.0 dq₁ dq₂ + 𝒪(‖x‖³)) t² + 𝒪(t³)

julia> z1N = zero(x1N)
  0.0 + 𝒪(‖x‖³) + 𝒪(t³)

julia> @btime TaylorSeries.mul!($z1N,$x1N,$y1N,0)
  83.679 ns (0 allocations: 0 bytes)
```

With this `zero!` methods enables doing things like:

```julia

julia> x1 = Taylor1(rand(3))
 0.6727385330826134 + 0.023575091119572567 t + 0.408205561614307 t² + 𝒪(t³)

julia> xN = sin(dq[1]+dq[2])
 1.0 dq₁ + 1.0 dq₂ + 𝒪(‖x‖³)

julia> @btime TaylorSeries.zero!($x1)
  8.049 ns (0 allocations: 0 bytes)

julia> @btime TaylorSeries.zero!($xN)
  21.815 ns (0 allocations: 0 bytes)

julia> @btime TaylorSeries.zero!($x1N)
  63.053 ns (0 allocations: 0 bytes)
```
without a single allocation.

This is still very much WIP, but do you think this is an idea worth pursuing @lbenet?