@lbenet 2024-01-31
I think it is worth pursuing indeed, since the results are really good! 

Once said that, let me point to [this line](https://github.com/JuliaDiff/TaylorSeries.jl/blob/f5587a772e99d93428c5201f5422d5629fa4fcaf/src/functions.jl#L244), which defines `zero!(res, a, ord)`, and supposedly avoids allocations; I'm not sure why we are incurring into allocating there...