@lbenet 2024-01-31
We should indeed keep the methods, e.g. for initialization of the aux quantities created by `@taylorize` in TaylorIntegration!