@lbenet 2024-02-01
( I guess a better place for those functions would be in `src/functions.jl`...) The idea is truly nice! Thanks a lot!