@PerezHz 2024-02-02
I'm trying to also avoid allocations in `TaylorSeries.div!`, and around this line I found there's an auxiliary `TaylorN` being declared there, which allocates once for each order of expansion:
https://github.com/JuliaDiff/TaylorSeries.jl/blob/f5587a772e99d93428c5201f5422d5629fa4fcaf/src/arithmetic.jl#L763
I'm thinking that a possible workaround is to add it as an argument to `div!`, so that that auxiliary can be declared somewhere else, but that change would be perhaps a bit breaking? There are other mutating functions with similar auxiliaries defined in the function body, which could benefit from a similar treatment, what are your thoughts on this @lbenet?