@PerezHz 2024-02-02
In other words, whereas this PR makes `mul!` allocation-free for mixtures, `div!` is not:
```julia
julia> @btime TaylorSeries.mul!($z1N,$x1N,$y1N,0)
  55.880 ns (0 allocations: 0 bytes)

julia> @btime TaylorSeries.div!($z1N,$x1N,$y1N,0)
  360.201 ns (15 allocations: 1.09 KiB)
```