@lbenet 2024-02-02
> I'm trying to also avoid allocations in `TaylorSeries.div!`, and around this line I found there's an auxiliary `TaylorN` being declared there, which allocates once for each order of expansion:
> 
> https://github.com/JuliaDiff/TaylorSeries.jl/blob/f5587a772e99d93428c5201f5422d5629fa4fcaf/src/arithmetic.jl#L763
> 

I don't quite remember the actual reason why that `aux` is needed there, but I think it is related to the sum in the recurrence relation... I recall the dirty tricks needed when defining sums and using `@taylorize`...

> I'm thinking that a possible workaround is to add it as an argument to `div!`, so that that auxiliary can be declared somewhere else, but that change would be perhaps a bit breaking? There are other mutating functions with similar auxiliaries defined in the function body, which could benefit from a similar treatment, what are your thoughts on this @lbenet?

I don't think we should care about how breaking that change would be to the code; we simply release a new patch version 😄. Once said this, my guess is that such a change impacts in the infrastructure needed for `@taylorize` to work nicely. I suggest adding a couple of comments (on improving `div!`) and leave those changes to another PR. My guess is that such auxiliaries occur beyond `div!`, so solving `div!` will tell us the way to proceed.