@PerezHz 2024-02-08
A workaround is to simply broadcast `zero`:
```julia
z.coeffs[2:end] .= zero.(z.coeffs[1]) # <-- broadcasted zero
```