@PerezHz 2024-02-08
> I can't reproduce the problem you report https://github.com/JuliaDiff/TaylorSeries.jl/pull/347#issuecomment-1934890667... I'm using Julia v1.10, and TaylorSeries from master...

Sorry for the confusion, the issue happens with this PR's branch