@PerezHz 2024-02-09
> I can't reproduce the problem you report https://github.com/JuliaDiff/TaylorSeries.jl/pull/347#issuecomment-1934890667... I'm using Julia v1.10, and TaylorSeries from master...

... now, if you meant that the issue pops up only using this PR then yes, although I actually would say the issue is hidden in TaylorIntegration. I'll explain in the next comment.