@PerezHz 2024-02-09
Using current TS master:

```julia
julia> using TaylorSeries

julia> dq = set_variables("δq", order=1, numvars=2)
2-element Vector{TaylorN{Float64}}:
  1.0 δq₁ + 𝒪(‖x‖²)
  1.0 δq₂ + 𝒪(‖x‖²)

julia> x = Taylor1([1.0+dq[1],-dq[1], dq[2], dq[1]-dq[2],dq[1]])
  1.0 + 1.0 δq₁ + 𝒪(‖x‖²) + ( - 1.0 δq₁ + 𝒪(‖x‖²)) t + ( 1.0 δq₂ + 𝒪(‖x‖²)) t² + ( 1.0 δq₁ - 1.0 δq₂ + 𝒪(‖x‖²)) t³ + ( 1.0 δq₁ + 𝒪(‖x‖²)) t⁴ + 𝒪(t⁵)

julia> z = deepcopy(x)
  1.0 + 1.0 δq₁ + 𝒪(‖x‖²) + ( - 1.0 δq₁ + 𝒪(‖x‖²)) t + ( 1.0 δq₂ + 𝒪(‖x‖²)) t² + ( 1.0 δq₁ - 1.0 δq₂ + 𝒪(‖x‖²)) t³ + ( 1.0 δq₁ + 𝒪(‖x‖²)) t⁴ + 𝒪(t⁵)

julia> z.coeffs[1] = x.coeffs[1]^1.5
 1.0 + 1.5 δq₁ + 𝒪(‖x‖²)

julia> z.coeffs[2:end] .= zero(z.coeffs[1]) # no zero broadcasting here
4-element view(::Vector{TaylorN{Float64}}, 2:5) with eltype TaylorN{Float64}:
  0.0 + 𝒪(‖x‖²)
  0.0 + 𝒪(‖x‖²)
  0.0 + 𝒪(‖x‖²)
  0.0 + 𝒪(‖x‖²)

julia> z[1][0] = HomogeneousPolynomial(-4.0,0) # only one z coefficient is updated
 - 4.0

julia> z # but the change is copied everywhere!
  1.0 + 1.5 δq₁ + 𝒪(‖x‖²) + ( - 4.0 + 𝒪(‖x‖²)) t + ( - 4.0 + 𝒪(‖x‖²)) t² + ( - 4.0 + 𝒪(‖x‖²)) t³ + ( - 4.0 + 𝒪(‖x‖²)) t⁴ + 𝒪(t⁵)
```
but this can be easily solved if we broadcast zero while initially filling the coefficients of z:
```julia
julia> z.coeffs[2:end] .= zero.(z.coeffs[1]) # zero is broadcasted here
4-element view(::Vector{TaylorN{Float64}}, 2:5) with eltype TaylorN{Float64}:
  0.0 + 𝒪(‖x‖²)
  0.0 + 𝒪(‖x‖²)
  0.0 + 𝒪(‖x‖²)
  0.0 + 𝒪(‖x‖²)

julia> z[1][0] = HomogeneousPolynomial(-4.0,0)
 - 4.0

julia> z
  1.0 + 1.5 δq₁ + 𝒪(‖x‖²) + ( - 4.0 + 𝒪(‖x‖²)) t + 𝒪(t⁵)
```
AFAIU, the issue is that if `zero` is not broadcasted then each of the elements of `z.coeffs[2:end]` point to the same place in memory:
```julia
julia> z.coeffs[2:end] .= zero(z.coeffs[1]) # no zero broadcast
4-element view(::Vector{TaylorN{Float64}}, 2:5) with eltype TaylorN{Float64}:
  0.0 + 𝒪(‖x‖²)
  0.0 + 𝒪(‖x‖²)
  0.0 + 𝒪(‖x‖²)
  0.0 + 𝒪(‖x‖²)

julia> z.coeffs[2] == z.coeffs[3]
true

julia> z.coeffs[2] === z.coeffs[3] # check if memory address is the same
true

julia> z.coeffs[2:end] .= zero.(z.coeffs[1]) # now, broadcast zero
4-element view(::Vector{TaylorN{Float64}}, 2:5) with eltype TaylorN{Float64}:
  0.0 + 𝒪(‖x‖²)
  0.0 + 𝒪(‖x‖²)
  0.0 + 𝒪(‖x‖²)
  0.0 + 𝒪(‖x‖²)

julia> z.coeffs[2] == z.coeffs[3]
true

julia> z.coeffs[2] === z.coeffs[3] # now each coefficient points to a different place in memory
false
```