@PerezHz 2024-02-09
Now, this raises the question: why does this appear in this PR and not in current `master`? The line that makes the trick is this:
https://github.com/JuliaDiff/TaylorSeries.jl/blob/f5587a772e99d93428c5201f5422d5629fa4fcaf/src/power.jl#L279
the way it currently behaves it allocates a new place in memory of each of the elements of `res.coeffs`, but these lines are precisely the ones that this PR touches, updating coefficients inplace while avoiding allocations. Of course, as you mentioned earlier, there are other allocations that are more tricky to remove, but we can go step by step.

If downstream tests pass (TI, PE, NEOs) then I would propose to go ahead with this and make the update in TaylorIntegration, unless there are other considerations to be made?