@lbenet 2024-02-09
I think that `z.coeffs[2:end] .= zero(z.coeffs[1])` is (in this PR) somehow behaving as a view, while before (e.g. master) it was behaving as a copy; hence the allocations. Did you check that the workaround, i.e., apply broadcasting to `zero` too, does allocate?